+++ NEWS.md
@@ -10,6 +10,7 @@
 - Update the fingerprint rules for CONCEPT, TEMPORAL [MODLD-534](https://folio-org.atlassian.net/browse/MODLD-534)
 - Update IRIs with 'marc' to use 'library' instead [MODLD-792](https://folio-org.atlassian.net/browse/MODLD-792)
 - Introduce Hub fingerprinting [MODLD-870](https://folio-org.atlassian.net/browse/MODLD-870)
+- Update the fingerprint rule of ID_LOCAL to use NAME property & STATUS edge.
 
 ## 1.0.1 (03-12-2025)
 - Initial release